# openapi/customer.oas.yaml
openapi: 3.0.3
info:
  title: Customer Service API
  version: 1.0.0

paths:
  /customers:
    get:
      operationId: listCustomers
      summary: List all customers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [customers]
                properties:
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
    post:
      operationId: createCustomer
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomer'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /customers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getCustomer
      summary: Get a customer by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteCustomer
      summary: Delete a customer
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Customer:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string

    CreateCustomer:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string

  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string